Allocate IPC shared memory at runtime
The old system used a hardcoded address that corresponded to uncached
memory in DS, but not in DSi. This caused weird issues due to outdated
values being read from the ARM9, and nothing being sent to the ARM7.

This patch allocates this memory in main RAM in an arbitrary location
and uses the helper memUncached() of libnds so that it works in both DS
and DSi modes.