aslib: Fix timer setup helper
This helper was resetting TIMER1_DATA to 0 whenever the frequency changed.
This means that whenever the user changed the playback rate this timer
would be set to 0. The problem is that this value is used to determine how
many samples need to be decoded during the next iteration. The number of
samples required is the difference between the old value and the new value.
If the new value is 0, the difference will be very big, so there will be a
buffer overflow when decoding too many samples and the game will likely
crash.