aslib: Refactor timer handling on ARM7
It looks like DeSmuMe resets the value of the timer whenever you write to
TIMERx_CR. Unfortunately, we need the value to be preserved when only
the playback rate is changed. To simplify timer handling, this commit
splits AS_SetTimer() into three functions that are more explicit about
what they are used for: AS_{Start/Set/Stop}_Timer()

Thanks to this, AS_SetTimer() can be reduced to the bare minimum so that it
works everywhere.